去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21375 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnDataSet.cpp
@@ -113,7 +113,7 @@ namespace engine
 
       /// has data
       if ( SDB_OK == rc &&
-           pReply->header.messageLength > sizeof( MsgOpReply ) )
+           pReply->header.messageLength > (INT32)sizeof( MsgOpReply ) )
       {
          if ( !ownned )
          {
